pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hearth-memory"
version = "0.1.0"
description = "Local-first persistent AI memory system exposed as an MCP server"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Ryan" }]
dependencies = [
    "apsw>=3.46",
    "sqlite-vec>=0.1.6",
    "mcp>=1.20",
    "click>=8.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
hearth = "hearth.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/hearth"]
